+++ CHANGELOG.md
@@ -3,6 +3,11 @@ CHANGELOG
 
 0.40.0
 ------
+- Added `zero` event that is triggered when there's no match
+  ```sh
+  # Reload the candidate list when there's no match
+  echo $RANDOM | fzf --bind 'zero:reload(echo $RANDOM)+clear-query' --height 3
+  ```
 - New actions
     - Added `track` action which makes fzf track the current item when the
       search result is updated. If the user manually moves the cursor, or the